feat: Move `GeneralSettings` to kmpuiviews and Update Settings
This commit moves the `GeneralSettings` UI component to the `kmpuiviews` module, updates the `GeneralSettings` and related settings to use the new location, and makes various code cleanup tasks.

-   **`GeneralSettings`:**
    - Moved `GeneralSettings` to `kmpuiviews`.
    - Updated `GeneralSettings` in `UIViews` to use the new `GeneralSettings` from `kmpuiviews`.
    - Moved `ThemeSetting`, `AmoledModeSetting`, `ExpressivenessSetting`, `BlurSetting`, `PaletteSetting`, `GridTypeSettings`, `ShareChapterSettings`, `DetailPaneSettings`, and `ShowDownloadSettings` to kmpuiviews.
    - Removed the old settings from `UIViews`.
    - Added `historySettings` and `navigationBarSettings` to the `GeneralSettings` to help compose the settings
-   -  Moved `BlurSetting`, `ShareChapterSettings`, `ShowDownloadSettings` to kmpuiviews.
+   - Updated `BlurSetting`, `ShareChapterSettings`, `ShowDownloadSettings` to use the new kmp locations.
    - Updated `OnboardingContent` to use the new `BlurSetting`, `ShareChapterSettings`, and `ShowDownloadSettings` from `kmpuiviews`.
 - **Other**
-   - General cleanup.
-  - Removed toast and app update checks from `SettingsActivity` for tv.
+    - Removed toast and app update checks from `SettingsActivity` for tv.
+    - General cleanup.
